Переформирован компонент списка комментариев
