+++ project/src/components/comments-list/comments-list.tsx
@@ -1,46 +1,13 @@
-import {useEffect} from 'react';
-import {connect, ConnectedProps} from 'react-redux';
-
-import {State} from '../../types/state';
-import {ThunkAppDispatch} from '../../types/action';
-
 import CommentMessage from '../comment-message/comment-message';
+import {Reviews} from '../../types/review';
 
-import {fetchOfferComments} from '../../store/actions/api-actions';
-
-
-function mapStateToProps({commentsList, id}: State) {
-  return ({
-    commentsList,
-    id,
-  });
-}
-
-function mapDispatchToProps(dispatch : ThunkAppDispatch) {
-  return ({
-    getData(id: number) {
-      dispatch(fetchOfferComments(id));
-    },
-  });
-}
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
 
 type CommentsListProps = {
-  onCommentsCount: (length: number) => void;
+  commentsList: Reviews,
 }
-type PropsFromRedux = ConnectedProps<typeof connector>;
-type ConnectedComponentProps = CommentsListProps & PropsFromRedux;
 
 
-function CommentsList({commentsList, getData, id, onCommentsCount}: ConnectedComponentProps ):JSX.Element {
-
-  useEffect(() => {
-    getData(id);
-  }, [id]);
-
-  onCommentsCount(commentsList.length);
-
+function CommentsList({commentsList}: CommentsListProps  ):JSX.Element {
   return (
     <ul className="reviews__list">
       {
@@ -55,5 +22,5 @@ function CommentsList({commentsList, getData, id, onCommentsCount}: ConnectedCom
   );
 }
 
-export {CommentsList};
-export default connector(CommentsList);
+export default CommentsList;
+
